requirements.txt
python-dotenv
numpy
asyncio
pyaudio
g711
pydub
sounddevice